Add support for custom beatLists in songs
Previously, songs could only change bpm every 4 beats. This lets you
change the bpm whenever you want.

Extend the json format of the song to have an optional beatList field.
This is an array of [songTime, kind].
`songTime` is the time of the beat in ms (float).
`kind` is the type of beat. 0: Minor. 1: Major.
The camera zoom effect happens on Major beats. So the beatList lets you
zoom on whatever beat you want.